requirements/base.txt
django-environ==0.12.0
psycopg2-binary==2.9.9
django==6.0.1
djangorestframework==3.15.1
channels==4.1.0
channels-redis==4.2.0
djangorestframework-simplejwt==5.3.1
django-filter==24.2
celery==5.4.0
redis==5.0.4
django-modeltranslation==0.18.12
akshare
pandas==2.2.2
django-celery-beat
numpy
TA-Lib
